+++ dataIngestion/config.py
@@ -1,0 +1,1 @@
+LOCATION_ID = 'DRFZS882NW8FK'
