--- routes/squareRoutes/config.py
@@ -1,1 +1,0 @@
-LOCATION_ID = 'LVTH1QZM3G4HT'
